Improves fetching so that it is done in app.vue with a composable instead of having it in a middleware.
Removes old middleware fetch.
Adds fetching composable.
Changes so that storyblok uses useAsyncStoryblok instead of custom solution.
