+++ config/i18n.config.ts
@@ -11,12 +11,12 @@ export default {
     },
   ],
   defaultLocale: 'sv',
-  detectBrowserLanguage: {
+  detectBrowserLanguage: false, /* {
     useCookie: true,
     cookieKey: 'lang',
     alwaysRedirect: true,
     fallbackLocale: 'en',
-  },
+  }, */
   // differentDomains: true, // DifferentDomains are not supported in nuxt3.
   strategy: 'prefix_and_default',
   lazy: true, // Lazy is not fully supported in nuxt3. Changing language is not working with cookie.
